Add files via upload
This project is a machine learning and data analysis project that addresses the problem of identifying fake and spam accounts on Instagram. It uses Python to analyze profile features like profile pictures and follower-to-following ratios. A Random Forest model was built and achieved high accuracy, providing a successful method to detect spammers.